test: restructure folder to keep tests in a file
